+++ ui/common.py
